Added typecast in write()
FIXED: 
overloaded function write() has problem on some platforms.
Added typecast (uint8_t) to select the right function in case multiple ones exist.

+++ src/Tlv493d.cpp
@@ -284,7 +284,7 @@ void Tlv493d::resetSensor(uint8_t adr)     // Recovery & Reset - this can be han
 		mInterface.bus->write(0xFF);
 	} else {
 		// if the sensor shall be initialized with address 0x5E
-		mInterface.bus->write(0x00);
+		mInterface.bus->write((uint8_t)0x00);
 	}
 
 	mInterface.bus->endTransmission(true);
